+++ compiler-tests/src/test/kotlin/com/skydoves/compose/stability/compiler/tests/StabilityTestConfigurator.kt
@@ -62,7 +62,6 @@ class StabilityTestConfigurator(testServices: TestServices) :
     IrGenerationExtension.registerExtension(
       StabilityAnalyzerIrGenerationExtension(
         stabilityOutputDir = "",
-        projectDependencies = "",
         traceAll = traceAll,
         stabilityConfigurationFiles = stabilityConfigurationFiles,
       ),